Move UnexpectedTypeException from ResourceBundle into Component to remove broken dependencies
