+++ src/Sylius/Bundle/CoreBundle/EventListener/ImageUploadListener.php
@@ -11,10 +11,10 @@
 
 namespace Sylius\Bundle\CoreBundle\EventListener;
 
-use Sylius\Bundle\ResourceBundle\Exception\UnexpectedTypeException;
 use Sylius\Component\Core\Model\ProductInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
 use Sylius\Component\Core\Uploader\ImageUploaderInterface;
+use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 use Sylius\Component\Taxonomy\Model\TaxonInterface;
 use Sylius\Component\Taxonomy\Model\TaxonomyInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
@@ -59,7 +59,6 @@ class ImageUploadListener
         if ($subject->hasFile()) {
             $this->uploader->upload($subject);
         }
-
     }
 
     public function uploadTaxonomyImage(GenericEvent $event)
@@ -76,6 +75,5 @@ class ImageUploadListener
         if ($subject->getRoot()->hasFile()) {
             $this->uploader->upload($subject->getRoot());
         }
-
     }
 }
