+++ src/Sylius/Component/Core/Promotion/Checker/UserLoyalityRuleChecker.php
@@ -11,10 +11,10 @@
 
 namespace Sylius\Component\Core\Promotion\Checker;
 
-use Sylius\Bundle\ResourceBundle\Exception\UnexpectedTypeException;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Promotion\Checker\RuleCheckerInterface;
 use Sylius\Component\Promotion\Model\PromotionSubjectInterface;
+use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 
 /**
  * Checks if user is created before/after configured period of time.
