+++ src/Sylius/Component/Shipping/Processor/ShipmentProcessor.php
@@ -12,9 +12,9 @@
 namespace Sylius\Component\Shipping\Processor;
 
 use Doctrine\Common\Collections\Collection;
-use Sylius\Bundle\ResourceBundle\Exception\UnexpectedTypeException;
 use Sylius\Component\Shipping\Model\ShipmentInterface;
 use Sylius\Component\Shipping\Model\ShipmentItemInterface;
+use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 
 /**
  * Shipment processor.
